identity-verification-python/requirements.txt
autopep8==2.0.1
azure-functions==1.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
pycodestyle==2.10.0
requests==2.28.1
tomli==2.0.1
urllib3==1.26.14